+++ cli/index.js
@@ -11,12 +11,6 @@ program
     .alias('i')
     .description('Add boilerplate files and scripts to your project')
     .action(function(dir = process.cwd()) {
-        const pkg = path.resolve(dir, 'package.json');
-        if (!fs.existsSync(pkg)) {
-            printPackageWarning(pkg);
-            process.exit(1);
-        }
-
         prompt(install.questions).then(options => install(dir, options));
     });
 
@@ -40,11 +34,3 @@ program
     });
 
 program.parse(process.argv);
-
-function printPackageWarning(pkg) {
-    console.warn(`\x1b[31m
-    ${pkg} not found
- 
-    Please initialize first using "npm init" or "yarn init".
-    `);
-}