+++ cli/install.js
@@ -2,14 +2,25 @@ const path = require('path');
 const fs = require('fs-extra');
 const replace = require('replace-in-file');
 
+const ownName = require('../package.json').name;
+
 const supportedFiles = {
     babel: '.babelrc',
-    editorconfig: '.editorconfig',
     eslint: '.eslintrc.js',
-    esdoc: '.esdoc.js',
     prettier: 'prettier.config.js',
+    gitattributes: '.gitattributes',
+    editorconfig: '.editorconfig',
+    esdoc: '.esdoc.js',
     sublime: 'project.sublime-project'
 };
+
+const checkedFiles = [
+    supportedFiles.babel,
+    supportedFiles.eslint,
+    supportedFiles.gitattributes,
+    supportedFiles.editorconfig,
+    supportedFiles.prettier
+];
 
 const supportedFilesExtended = {
     babel: '.babelrc',
@@ -20,40 +31,76 @@ const supportedFilesExtended = {
 
 const supportedScripts = {
     build: 'babel src --out-dir lib --copy-files',
+    watch: 'yarn build --watch',
     lint: 'eslint src',
     docs: 'esdoc'
 };
+const checkedScripts = [supportedScripts.build, supportedScripts.watch, supportedScripts.lint];
 
 const DependencyMode = {
     NONE: 'no',
-    INDIRECT: 'yes - using only this package as a dependency',
-    DIRECT: 'yes - using all required packages as separate dependencies'
+    INDIRECT: `yes, indirectly: add only ${ownName} to devDependencies`,
+    DIRECT: 'yes, directly - add each required package to devDependencies'
+};
+
+const PackageManager = {
+    YARN: 'yarn',
+    NPM: 'npm'
+};
+const InstallMode = {
+    YES: 'yes',
+    NO: 'no'
 };
 
 const questions = [
     {
-        message: 'Choose config files to create',
+        message: 'Which package manager should we use?',
+        type: 'list',
+        name: 'manager',
+        default: PackageManager.NPM,
+        choices: [PackageManager.NPM, PackageManager.YARN]
+    },
+    {
+        message: 'Which config files should we create?',
         type: 'checkbox',
         name: 'files',
-        choices: Object.keys(supportedFiles).map(file => ({ name: file, checked: true }))
+        choices: Object.entries(supportedFiles).map(([key, value]) => ({
+            name: key,
+            checked: checkedFiles.includes(value)
+        }))
     },
     {
-        message: 'Choose scripts to add to package.json?',
+        message: 'Which scripts should we add to the package.json?',
         type: 'checkbox',
         name: 'scripts',
-        choices: Object.keys(supportedScripts).map(script => ({ name: script, checked: true }))
+        choices: Object.entries(supportedScripts).map(([key, value]) => ({
+            name: key,
+            checked: checkedScripts.includes(value)
+        }))
     },
     {
-        message: 'Add dependencies?',
+        message: 'Should we add the required devDependencies to the package.json?',
         type: 'list',
         name: 'dependencies',
         default: DependencyMode.INDIRECT,
-        choices: [DependencyMode.NONE, DependencyMode.DIRECT, DependencyMode.INDIRECT]
+        choices: [DependencyMode.INDIRECT, DependencyMode.DIRECT, DependencyMode.NONE]
+    },
+    {
+        message: 'Should we install the dependencies after init?',
+        type: 'list',
+        name: 'installDependencies',
+        default: InstallMode.YES,
+        choices: [InstallMode.YES, InstallMode.NO]
     }
 ];
 
-async function install(dir, { files, scripts, dependencies }) {
+async function install(dir, { manager, files, scripts, dependencies, installDependencies }) {
     console.log('\n[b7] install');
+
+    const pkg = path.resolve(dir, 'package.json');
+    if (!fs.existsSync(pkg)) {
+        await initializePackage(dir, manager);
+    }
 
     if (files && files.length) {
         // keys to values
@@ -66,7 +113,20 @@ async function install(dir, { files, scripts, dependencies }) {
     }
 
     if (dependencies !== DependencyMode.NONE) {
-        await installDependencies(dir, dependencies);
+        await addDependencies(dir, dependencies);
+    }
+
+    if (installDependencies !== InstallMode.NO) {
+        await installPackages(dir, manager);
+    }
+}
+
+async function initializePackage(dir, manager) {
+    try {
+        console.info('Initialize package');
+        await execAsync(`cd ${dir} && ${manager} init -y`);
+    } catch (error) {
+        console.error('\nFailed initializing package', error);
     }
 }
 
@@ -92,7 +152,7 @@ async function installFiles(dir, files, dependencyMode) {
 
         // perform copy operation
         try {
-            console.log(`[b7] copy file - "${source}" -> "${target}"`);
+            console.log(`[b7] add file - "${target}"`);
             await fs.copy(source, target);
         } catch (error) {
             console.error(`\nFailed copying file ${source} to ${target}`);
@@ -130,7 +190,16 @@ async function installScripts(dir, scripts) {
     }
 }
 
-async function installDependencies(dir, mode) {
+async function installPackages(dir, manager) {
+    try {
+        console.info('Installing packages');
+        await execAsync(`cd ${dir} && ${manager} install`);
+    } catch (error) {
+        console.error('\nFailed installing packages', error);
+    }
+}
+
+async function addDependencies(dir, mode) {
     try {
         const ownPackage = getPackage(path.resolve(__dirname, '..'));
         const targetPackage = getPackage(dir);
@@ -143,8 +212,6 @@ async function installDependencies(dir, mode) {
                 targetPackage.devDependencies[name] = version;
             });
             await fs.writeJson(path.resolve(dir, 'package.json'), targetPackage, { spaces: 2 });
-
-            console.log('\nPlease install the dependencies by executing "npm install" or "yarn install"');
         } else {
             console.log('\nNo missing dependencies found');
         }
@@ -174,10 +241,12 @@ function getMissingDependencies(ownPackage, targetPackage, mode) {
     switch (mode) {
         case DependencyMode.DIRECT:
             // install each dependency of this package as a dependency of the target package
-            return ownDeps.filter(name => targetDeps.indexOf(name) === -1).map(name => ({
-                name: name,
-                version: ownPackage.dependencies[name]
-            }));
+            return ownDeps
+                .filter(name => targetDeps.indexOf(name) === -1)
+                .map(name => ({
+                    name: name,
+                    version: ownPackage.dependencies[name]
+                }));
         case DependencyMode.INDIRECT:
             // install only this package as a dependency
             if (targetDeps.indexOf(ownPackage.name) > -1) {
@@ -201,3 +270,21 @@ module.exports = install;
 module.exports.questions = questions;
 module.exports.supportedFiles = supportedFiles;
 module.exports.supportedScripts = supportedScripts;
+
+/**
+ * Executes a shell command and return it as a Promise.
+ * @param cmd {string}
+ * @return {Promise<string>}
+ * @see https://medium.com/@ali.dev/how-to-use-promise-with-exec-in-node-js-a39c4d7bbf77
+ */
+function execAsync(cmd) {
+    const exec = require('child_process').exec;
+    return new Promise((resolve, reject) => {
+        exec(cmd, (error, stdout, stderr) => {
+            if (error) {
+                console.error('[execAsync] failed', error);
+            }
+            resolve(stdout ? stdout : stderr);
+        });
+    });
+}