+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# 1.0.1
+# 1.2.0
+
+-   automatically initialize package when `package.json` is missing
+-   add `npm` or `yarn` choice
+-   add `.gitattributes` config
+-   add option to install the dependencies
+
+# 1.1.0
 
 -   Fix dependencies used by CRA to match `react-scripts@2.1.1`
 -   Add `@babel/plugin-proposal-export-default-from`